Add fork of max-statements-per-line
The base ESLint max-statements-per-line rule gives false positives if
you use an expression inside the body of a one line arrow function.

For example:

```
myArray.forEach((x) => { y += x});
```

The above line will be counted as 2 statements.

Since this behavior was not desirable, I have forked the
max-statements-per-line rule and adjusted it so that it does not count the
one line arrow function in the total statements on that line.