fix: remove forked indent rule (#41)
The reason for forking the indent rule was to allow us to line up
variable declarations. The original indent rule only allowed doing this
if you used 2 or 4 space indent. The new eslint rule has a setting to
require variable declarations to line up, so this rule is no longer
needed. In fact, our own config now uses the default rule.