refactor: safely require ast-utils
The location of the ast-utils module changed in the upgrade from eslint
v5 to eslint v6. Because we are requiring this module for our forked
rules from eslint core, this prevented us from upgrading to eslint v6.

This commit adds a safe require for the module in order to make the
plugin compatible with either version of eslint.